requirements.txt
altgraph==0.17.3
Pillow==9.3.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
